requirements.txt
daytona
google-generativeai